Loading bug fixed
Sometimes when you click the dropdown/menu item in the list/click away when the list is open.
Then you click the dropdown again, it either doesn't open or it opens halfway and closes right away.

this also happens when you keep clicking repeatedly fast the dropdown.

The root cause of this is the animation, the slideup/slidedown/settimeout functions. I'm currently in a rush and have not enough time to read the whole code and find all slides and timeouts to fix them properly.

So i added a temporary fix, see line 95 to line 102. 
This prevents repeatedly fast clicks interaction with the button, which causes the dropdown animation to bug from too many animation queue.
I set the removing of the loading class to 600 millis to match the other timeouts/slides, 100ms, 200ms and 600 ms.
This makes sure all the animation finishes before the user can interact again with the dropdown button. 

Also added some missing semi-colons on the following lines:
line 151
line 247

to fix the animation with slides, you have to add the property done e.g.

```
form_element.slideDown({
    start: function()
    {
        button.addClass("is-active");
        button.children(".fas").addClass("fa-rotate-180");
    },
    done: function()
    {
        button.prop("disabled", false);
    }
});
```